fix: unblock deploy — client-bundle leak in roster admin, invalid S3 lifecycle rule
Two independent deploy blockers on dev, both from today's merges, neither
catchable by CI as it stands.

1) FRONTEND IMAGE COULD NOT BUILD — server code in the client bundle.

   `next build` failed with "Can't resolve 'fs'", "Can't resolve 'net'", and
   UnhandledSchemeError on node:async_hooks / node:crypto. Import trace:

     app/(protected)/admin/rosters/_components/rosters-admin.tsx  ("use client")
       -> lib/roster/status.ts
       -> lib/db/drizzle-client.ts
       -> lib/logger.ts
       -> winston -> fs / tail-file

   The component's two `import type` lines are erased at compile and were
   never the problem. The VALUE import of `isOneRosterSyncStatusActive` is
   what pulls all of status.ts — and therefore the Drizzle client — into the
   browser bundle.

   Split the pure half (zod schemas, ONEROSTER_SYNC_ACTIVE_WINDOW_MS,
   isOneRosterSyncStatusActive, parseOneRosterSyncStatus — no I/O) into
   lib/roster/status-shared.ts. status.ts keeps the DB functions and
   re-exports the shared surface, so every existing server-side import
   continues to work unchanged. The client component now imports from
   status-shared.

2) AGENT PLATFORM STACK UPDATE FAILED — illegal S3 lifecycle rule.

   CloudFormation: "AbortIncompleteMultipartUpload cannot be specified with
   Tags." The `cleanup-private-workspace-versions` rule carried both
   `tagFilters: { Scope: 'private' }` and
   `abortIncompleteMultipartUploadAfter`. S3 forbids that pairing because an
   in-flight multipart upload has no object tags yet — tags are applied on
   completion — so a tag-filtered rule could never match one.

   Removed the action from the tag-filtered rule and added a separate
   `abort-abandoned-multipart-uploads` rule with no prefix or tag filter,
   which is a legal home for it and preserves the cleanup intent.

VERIFICATION (both reproduced and re-run, not reasoned about):
  - `next build --webpack`: "✓ Compiled successfully in 58s", exit 0. All
    four original error classes now count zero.
  - `cdk synth AIStudio-AgentPlatformStack-Dev`: exit 0. Generated template
    shows the tag-filtered rule WITHOUT AbortIncompleteMultipartUpload and a
    separate unfiltered rule carrying it.
  - tsc clean; tests/unit/oneroster-admin-settings.test.ts 7/7 (it imports
    isOneRosterSyncStatusActive and still resolves through the re-export).

Note for anyone reproducing locally: `bun install` may be needed first —
@aws-sdk/client-scheduler is declared in package.json but was missing from a
stale node_modules, which masks itself as an unrelated build failure. The
Docker build does a frozen-lockfile install and is unaffected.
